backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.48
alembic==1.13.1
psycopg2-binary==2.9.10

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation and utilities
pydantic==2.10.5
pydantic-settings==2.7.1
email-validator==2.2.0

# HTTP client for Ollama
httpx==0.26.0

# Background jobs (optional, for scheduled tasks)
apscheduler==3.10.4

# CORS
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code quality
black==23.12.1
ruff==0.1.11
mypy==1.8.0

# Date utilities
python-dateutil==2.8.2
